--- postnet/commands/evalInput.js
@@ -1,7 +1,0 @@
-"use strict";
-
-module.exports = function () {
-    return {
-        text:`please give right input:`
-    }
-};
