+++ postnet/commands/go-eval-input-command.js
@@ -1,0 +1,12 @@
+"use strict";
+let CommandResponse = require('../src/command-response');
+
+class GoEvalInputCommand {
+    run() {
+        return new CommandResponse({
+            error:`please give right input`
+        })
+    }
+}
+
+module.exports = GoEvalInputCommand;
