+++ postnet/commands/go-exit-command.js
@@ -1,0 +1,10 @@
+
+"use strict";
+
+class GoExitCommand {
+    run(){
+        process.exit();
+    }
+}
+
+module.exports = GoExitCommand;
