+++ postnet/commands/go-to-zipcode-command.js
@@ -1,0 +1,15 @@
+"use strict";
+
+let ZipcodeToBarcodeCommand = require('./zipcode-to-barcode-command');
+let CommandResponse = require('../src/command-response');
+
+class GoToZipcodeCommand {
+    run(){
+        return new CommandResponse({
+            text:`please input zip code:`,
+            newMapping:{"*":new ZipcodeToBarcodeCommand()}
+        })
+    }
+}
+
+module.exports = GoToZipcodeCommand;
