--- postnet/commands/goToZipCode.js
@@ -1,9 +1,0 @@
-let translateZipCodeToBarcode = require('./translateZipCodeToBarcode');
-
-module.exports = function () {
-    return {
-        text:`please input zip code:`,
-        newMapping:{"*":translateZipCodeToBarcode}
-    }
-};
-
