+++ postnet/src/route-response.js
@@ -1,0 +1,9 @@
+
+class RouteResponse {
+    constructor({text,rerun}){
+        this.text = text;
+        this.rerun = rerun;
+    }
+}
+
+module.exports = RouteResponse;
